consolidate related classes
